// packages/@fluentui-vue/next/src/components/Rating/Rating.types.ts
export interface IRatingStarProps {
  fillPercentage: number
  disabled?: boolean
  classNames: IProcessedStyleSet<IRatingStyles>
  icon: string
  starNum?: number
  unselectedIcon?: string
}
// ...
export enum RatingSize {
  Small = 0,
  Large = 1,
}

/**
 * {@docCategory Rating}
 */
export interface IRatingStyleProps {
  disabled?: boolean
  readOnly?: boolean
  theme: ITheme
}

/**
 * {@docCategory Rating}
 */
export interface IRatingStyles {
  root: IStyle
  ratingStar: IStyle
  ratingStarBack: IStyle
  ratingStarFront: IStyle
  ratingButton: IStyle
  ratingStarIsSmall: IStyle
  ratingStarIsLarge: IStyle
  rootIsSmall: IStyle
  rootIsLarge: IStyle
  labelText: IStyle
  ratingFocusZone: IStyle
}